airflow_pyspark_iceberg/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-trino==5.7.0
apache-airflow-providers-snowflake
apache-airflow-providers-apache-spark==4.11.3
apache-airflow-providers-common-compat==1.2.2
boto3
python-dotenv
requests
dbt-core==1.9.0b3
dbt-snowflake==1.8.4
sqlfluff==3.0.6
astronomer-cosmos
pyiceberg[s3fs,pyarrow]
